Resolve conflict with func_802300F0

+++ src/kernel/code_30EA0.h
@@ -6,6 +6,6 @@
 
 void func_8022FEF0(UnkStruct_80204D94* arg0, u16 arg1, uvUnkTileStruct* arg2, Unk80225FBC_0x28* arg3, u16 arg4, u8 arg5);
 void func_802300D0(Unk80225FBC_0x28* arg0);
-s32 func_802300F0(u16);
+u16 func_802300F0(s32);
 
 #endif // KERNEL_CODE_30EA0_H
